+++ apps/client/src/views/public/Public.scss
@@ -1,4 +1,4 @@
-@use '../../../theme/viewerDefs' as *;
+@use '../../theme/viewerDefs' as *;
 
 .public-screen {
   margin: 0;
